[기능] Memoir 삭제, 조회 API 구현중2
